+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2018-11-08
+### Added
+- Lexeme locale and spellings
+- Japanese expander
+
 ## [0.3.1] - 2018-11-05
 ### Updated
 - Lexicon.toFile now returns promise